agent_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
paramiko==3.3.1
requests==2.31.0
aiohttp==3.9.1
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-socketio==5.10.0 